Add mouse object to an experiment

This function takes an experiment object and mouse object and adds the mouse object to the experiment. The mouse's unprocessed data, including all of it's individual slice information and raw imported segmentation and registration data will be not be added from the mouse object to save space. Any desire to modify this data must be done at the mouse object level before continuing further.

This function will also read the individual mouse attributes and automatically populate the experimental attributes that are relevant. For example, the 'group' attribute of a mouse will be read and automatically added to the experiment object's 'experimental_groups' attribute if it is a new unique experimental group name.

Usage

add_mouse(e, m, replace = FALSE)

Arguments

e

experiment object

m

mouse object

replace

(bool, default = FALSE) Replace a mouse already contained in an experiment object.

Value

an experiment object

Examples

e <- experiment(experiment_name = "example")
m <- mouse(mouse_ID = "Mouse1")
e <- add_mouse(e, m)
e <- add_mouse(e, m, replace = TRUE)
#> There was existing data found for mouse Mouse1
#> Replaced existing mouse data!