Check for redundant parent regions included in a list of acronyms in a plate. For example, if all the the subregions for the hypothalamus are represented, the HY should not be included in the list.

Check for redundant parent regions included in a list of acronyms in a plate. For example, if all the the subregions for the hypothalamus are represented, the HY should not be included in the list.

Usage

check_redundant_parents(acronyms, ontology = "allen")

Arguments

acronyms

(vec) a vector of acronyms to check for possible parents that are redundantly included in the vector.

ontology

(str, default = "allen") Region ontology to use. options = "allen" or "unified"

Value

A list containing two elements: one vector of unique child acronyms, a vector of the parent regions considered redundant

Examples

check_redundant_parents(acronyms = c("ACA5", "ACA1", "ACA"))
#> $unique_acronyms
#> [1] "ACA5" "ACA1"
#> 
#> $redundant_parents
#> [1] "ACA"
#> 
check_redundant_parents(acronyms = c("VMHDM", "VMHSh", "VMH"), ontology = "unified")
#> $unique_acronyms
#> [1] "VMHDM" "VMHSh"
#> 
#> $redundant_parents
#> [1] "VMH"
#>