Create graph objects for plotting different analysis subgroups.

Create graph objects for plotting different analysis subgroups.

Usage

create_networks(
  e,
  correlation_list_name,
  channels = c("cfos", "eyfp", "colabel"),
  proportional_thresh = NULL,
  alpha = 0.05,
  pearson_thresh = 0,
  ontology = "allen",
  anatomical.order = c("Isocortex", "OLF", "HPF", "CTXsp", "CNU", "TH", "HY", "MB", "HB",
    "CB"),
  filter_isolates = FALSE
)

Arguments

e

experiment object

correlation_list_name

(str) Name of the correlation list object used to generate the networks.

channels

(str, default = c("cfos", "eyfp", "colabel")) The channels to process.

proportional_thresh

(float, default = NULL) Takes precedent over the alpha and the pearson_thresh parameters. Input the desired edge proportion (i.e., edge density) as your desired sparsity constraint.

alpha

(float, default = 0.05) The significance threshold for including brain regions in the network. if NULL or NA, this threshold is not applied.

pearson_thresh

(float, default = 0) The pearson correlation coefficient threshold to apply for filtering out

ontology

(str, default = "allen") Region ontology to use. options = "allen" or "unified"

anatomical.order

(vec, c("Isocortex","OLF","HPF","CTXsp","CNU","TH","HY","MB","HB","CB")) The default super region acronym list that groups all subregions in the dataset.

filter_isolates

(logical, default = FALSE) Whether to filter out the number of isolated (zero degree) nodes from the network. Default is to retain them.

Value

e experiment object. This object now has a new added element called networks. This is a list storing a graph object per channel for each network analysis run. The name of each network (network_name) is the same as the correlation_list_name

used to generate the network. This network_name is fed as a parameter into the plot_networks() function.

See also

Examples