Create an experiment object

experiment() constructs an S3 object of class 'wb_experiment'. An experiment object consists of a list of processed mouse objects with raw data from slices omitted, and experimental attributes stored as a list.

Usage

experiment(
  experiment_name = NULL,
  experimenters = NULL,
  channels = NULL,
  experiment_groups = NULL,
  drug_groups = NULL,
  sex_groups = NULL,
  cohorts = NULL,
  strains = NULL,
  genotypes = NULL,
  reporters = NULL,
  ages = NULL,
  output_path = "set output path for your experiment",
  ...
)

Arguments

experiment_name

(str, default = NULL)

experimenters

(str, default = NULL)

channels

(str, default = NULL) Autogenerated with add_mouse() function. Will detect all unique channels stored in a mouse object.

experiment_groups

(str, default = NULL) Autogenerated with add_mouse() function. Must exactly match the string values from the mouse objects.

drug_groups

(str, default = NULL) Autogenerated with add_mouse() function. Must exactly match the string values from the mouse objects.

sex_groups

(str, default = NULL) Autogenerated with add_mouse() function. Must exactly match the string values from the mouse objects.

cohorts

(str, default = NULL) Autogenerated with add_mouse() function. Must exactly match the string values from the mouse objects.

strains

(str, default = NULL) Autogenerated with add_mouse() function. Must exactly match the string values from the mouse objects.

genotypes

(str, default = NULL) Autogenerated with add_mouse() function. Must exactly match the string values from the mouse objects.

reporters

(str, default = NULL) Autogenerated with add_mouse() function. Must exactly match the string values from the mouse objects.

ages

(str, default = NULL). Autogenerated with add_mouse() function. Must exactly match the string values from the mouse objects.

output_path

(str, default = 'set output path for your experiment') Where to save the RData file for your experiment object

...

additional custom keyword pair attributes you'd like to store

Value

An experiment, a colloquial term for an object of class 'wb_experiment'. An 'experiment' object is also a list, with class list.

Details

The experimental attributes can be assigned as arguments to the experiment constructor function. See the parameters listed for the default values for these attributes Note that you are able to add custom attributes as keyword pairs, if you would like to keep track of an additional piece of information. However, this will only serve a descriptive purpose and will not be used for analysis. You may not need to use all experimental attributes but fill out as many are applicable to your experiment.

See also

Examples

my_experiment <- experiment() # constructs an experiment object