map_cells_to_atlas (generic function)

Method for forward warping segmentation data to atlas space for a slice object. Requires segmentation objects to be made for channels specified and a registration object.

Method for forward warping segmentation data to atlas space for a slice within a mouse object. Requires segmentation objects to be made for the channels specified and a registration.

Usage

map_cells_to_atlas(x, ...)

# S3 method for class 'slice'
map_cells_to_atlas(
  x,
  channels = NULL,
  clean = TRUE,
  display = TRUE,
  mouse_ID = NULL,
  ...
)

# S3 method for class 'mouse'
map_cells_to_atlas(
  x,
  slice_ID = NA,
  hemisphere = NULL,
  channels = NULL,
  clean = TRUE,
  display = TRUE,
  replace = FALSE,
  ...
)

Arguments

x

a mouse or slice object

...

additional parameters besides 'registration', 'segmentation', 'forward.warps', and 'device' to pass to the wholebrain::inspect.registration() function

channels

(str vector, default = NULL) Channels to process. If NULL, defaults to the channels stored in the slice object attributes.

clean

(bool, default = TRUE). Remove cells that don't map to any regions.

display

(bool, default = TRUE). Display the results of the forward warp for the slice.display

mouse_ID

(str) mouse ID

slice_ID

(str) ID of slice

hemisphere

(str) 'left', 'right' or NULL (both)

replace

(bool, default = FALSE). Replace current forward warped data, both raw and cleaned.

Value

a mouse or slice object

m mouse object

Examples