Normalize cell counts per mm^2^ or by mm^3^ (if multiplying by the stack size).

Run this function after all the slices that you want to process are finished being added and you have combined your cell counts with get_cell_table(). This functions process all channels where a cell table was made using the latter function.

Usage

normalize_cell_counts(
  m,
  combine_hemispheres = TRUE,
  simplify_regions = TRUE,
  simplify_keywords = c("layer", "part", "stratum", "division", "leaflet",
    "Subgeniculate", "island", "Islands", "Fields of Forel", "Cajal", "Darkschewitsch",
    "Precommissural"),
  split_hipp_DV = TRUE,
  DV_split_AP_thresh = -2.7
)

Arguments

m

mouse object

combine_hemispheres

(bool, default = TRUE) Combine normalized cell counts from both hemispheres

simplify_regions

(bool, default = TRUE ) simplify the normalized region counts based on keywords in the internal function, simplify_keywords

simplify_keywords

(str vec, default = c("layer","part","stratum","division")). Keywords to search through region names and simplify to parent structure

split_hipp_DV

(bool, default = TRUE) Split the subregions of the CA1, CA2, CA3, and DG based on a specified AP coordinate cutoff. This is because the Allen atlas doesn't have a dorsal/ventral region designation for these ROIs.

DV_split_AP_thresh

(numeric, default = -2,7) The specified AP coordinate threshold to split hippocampal cell counts into dorsal and ventral.

Value

m mouse object

Examples

if (FALSE) { # \dontrun{
m <- normalize_cell_counts(m, combine_hemispheres = TRUE, simplify_regions = TRUE)
} # }