Plot the degree distributions across regions

Bar plot of degree per region in descending magnitude

Usage

plot_degree_regions(
  e,
  channels = c("cfos", "eyfp"),
  colors = c("red", "green"),
  network = "AD",
  title = "",
  height = 10,
  width = 20,
  ylim = c(0, 15),
  sort_super_region = FALSE,
  region_label_angle = 60,
  label_text_size = 12,
  filter_isolates = TRUE,
  image_ext = ".png",
  print_plot = FALSE,
  save_plot = TRUE,
  theme.bar = NULL
)

Arguments

e

experiment object

channels

(str, default = c("cfos", "eyfp", "colabel")) Channels to plot

colors

(str, default = ) String vector of hexadecimal color codes corresponding to to each channel plotted.

network

(str, default = "AD") Which network to plot the degree distribution across regions

title

(str, default = "") Plot title.

height

(int, default = 15) Height of the plot in inches.

width

(int, default = 20) Width of the plot in inches.

ylim

(vec, default = c(0,15))axes limits of y-axis

sort_super_region

(bool, default = FALSE) Whether to divide into subfacets based on which parent region

region_label_angle

(int, default = 60) Angle of region labels.

label_text_size

(int, default = 12) Font size of region labels.

filter_isolates

(default = TRUE) Avoid plotting isolated nodes (zero value)

image_ext

(default = ".png") image extension to the plot as.

print_plot

(bool, default = FALSE) Whether to print the plot as an output.s

save_plot

(bool, default = TRUE) Save into the figures subdirectory of the the experiment object output folder.

theme.bar

User option to use their own ggplot theme the experiment object output folder.

Value

p_list A list the same length as the number of channels, with each element containing a plot handle for that channel.

Examples

if (FALSE) { # \dontrun{
p <- plot_degree_regions(e, colors = c("#660000", "#FF0000"),
channels = "cfos", region_label_angle = 60,
ylim = c(0, 15), image_ext = ".png")
} # }