Plot mean global efficiency

Plot the mean global efficiency of the networks in a barplot. Error bars are plotted as SEM.

Usage

plot_mean_global_effic(
  e,
  color_palettes = c("reds", "greens"),
  colors_manual = NULL,
  channels = c("cfos", "eyfp"),
  labels = c(AD = "AD_label", control = "control_label"),
  title = "my_title",
  height = 10,
  width = 10,
  label_angle = 60,
  rev_x_scale = FALSE,
  ylim = c(0, 0.7),
  theme.gg = NULL,
  image_ext = ".png",
  print_plot = FALSE,
  save_plot = TRUE
)

Arguments

e

experiment object

color_palettes

(str, default = c("reds", "greens")) Color palettes from grDevices::hcl.colors that are used to for plotting networks characteristics for each channel, respectively.

colors_manual

(str, default = NULL ) Manually choose the hexadecimal color codes to create a custom color palette, e.g. colors_manual = c("#660000", "#FF0000", "#FF6666"). Warning: this will be applied to all channels. It's recommended to set the channels parameter to a single channel if this parameter is used.

channels

(str, default = c("cfos", "eyfp", "colabel")) Channels to plot

labels

The labels to correspond with your network names.

title

(str, default = "my_title) plot title

height

(int, default = 10) Height of the plot in inches.

width

(int, default = 10) Width of the plot in inches.

label_angle

(int, default = 60)

rev_x_scale

(bool, default = FALSE) Reveres the scale of the categorical variables

ylim

(vec, default = c(0,10)) Axes limits of y-axis

theme.gg

(default = NULL) Option to use custom ggplot2 theme if the user wants

image_ext

(default = ".png") image extension to the plot as.

print_plot

(bool, default = FALSE) Whether to print the plot as an output.s

save_plot

(bool, default = TRUE) Save into the figures subdirectory of the the experiment object output folder.

Value

p_list A list the same length as the number of channels, with each element containing a plot handle for that channel.

Examples

if (FALSE) {
p <- plot_mean_global_effic(e, colors_manual = c("#660000", "#FF0000"),
channels = "cfos", labels = c("AD" = "AD_label", "control" = "control_label"),
title = "my title", ylim = c(0, 0.7), image_ext = ".png")
}