Summarize multiple networks. calculate network statistics for each network. This is not meant to summarize networks created using create_joined_networks.

Summarize multiple networks. calculate network statistics for each network. This is not meant to summarize networks created using create_joined_networks.

Usage

summarise_networks(
  e,
  network_names,
  channels = c("cfos", "eyfp", "colabel"),
  save_stats = TRUE,
  save_degree_distribution = TRUE,
  save_betweenness_distribution = TRUE,
  save_efficiency_distribution = TRUE
)

Arguments

e

experiment object

network_names

(str) The names of the networks to generate summary tables for, e.g. network_names = c("female_AD", "female_control")

channels

(str, default = c("cfos", "eyfp", "colabel")) The channels to process.

save_stats

(bool, default = TRUE) Save the summary stats as a csv file in the output folder. Note that the clustering calculated is an average of the local vertex clustering.

save_degree_distribution

(bool, default = TRUE) Save the network degree distributions (frequencies of each degree) across each comparison group as a csv file.

save_betweenness_distribution

(bool, default = TRUE) Save the betweenness distribution and summary as a csv.

save_efficiency_distribution

(bool, default = TRUE) Save the efficiency distribution and summary as a csv.

Value

e experiment object

Examples

if (FALSE) { # \dontrun{
e <- get_network_statistics(e,  network_names = c("female_AD", "female_control"),
channels = c("cfos", "eyfp", "colabel"), save_stats = TRUE, save_degree_distribution = TRUE)
} # }