Summarize the parameters of the rewired null networks generated by rewire_network()

Summarize the parameters of the rewired null networks generated by rewire_network()

Usage

summarize_null_networks(
  null_nodes_list,
  network_names = NULL,
  channel = "cfos"
)

Arguments

null_nodes_list

a list of output summary tables (1 per network) of rewired network properties of all nodes from rewire_network().

network_names

(str vec) Name of the networks that were rewired (in the same order as the list). null_nodes_list and network_names must have the same length.

channel

(str) channel to process

Value

a list of length 2. The first element is named global_summary and contains a table of global summary statistics. The second element is named node_summary, and contained per node statistics averaged from multiple null networks.

Examples

if (FALSE) { # \dontrun{
rewire_summary <- rewire_network(e, "Context", channels = "eyfp", return_graphs = FALSE)
summarized_null_networks <- summarize_null_networks(rewire_summary,
network_names = "Context", channel = "eyfp")
} # }