Plot the results of the permutation histogram used to determine the p-value of the pairwise region comparison

Create a Volcano plot.

Usage

volcano_plot(
  e,
  permutation_comparison = "female_AD_vs_male_AD",
  channels = c("cfos", "eyfp", "colabel"),
  colors = c("#be0000", "#00782e", "#f09b08"),
  save_plot = TRUE,
  title = NULL,
  ylim = c(0, 3),
  height = 8,
  width = 10,
  print_plot = FALSE,
  plt_theme = NULL,
  point_size = 1,
  image_ext = ".png"
)

Arguments

e

experiment object

permutation_comparison

The name of the correlation group comparisons to plot.

channels

(str, default = c("cfos", "eyfp", "colabel")) channels to plot.

colors

(str, default = c("#be0000", "#00782e", "#f09b08")) Hexadecimal code for the colors corresponding to the channels attribute of the correlation_list. Color values can also be input compatible with ggplot2 plotting functions.

save_plot

(bool, default = TRUE) Save into the figures subdirectory of the the experiment object output folder.

title

Title of the plot.

ylim

(vec, default = c(0,y)) Y-axis range (logarithmic).

height

height of the plot in inches.

width

width of the plot in inches.

print_plot

(bool, default = FALSE) Print the plot as graphics windows.

plt_theme

(default = NULL) Add a ggplot2::theme() to the plot. If NULL, the default is taken..

point_size

(default = 1) Size of the plotted points.

image_ext

(default = ".png") image extension to save the plot as.

Value

p_list A list the same length as the number of channels, with each element containing a plot handle for that channel.

Details

Plot the correlation difference between two comparison groups into a volcano plot. The function correlation_diff_permutation() must be run first in order to generate results to plot.

Examples

if (FALSE) { # \dontrun{
volcano_plot(e, permutation_comparison = "female_AD_vs_male_AD",
channels = c("cfos", "eyfp", "colabel"), colors =  c("#be0000", "#00782e", "#f09b08"),
save_plot = TRUE, title = NULL, ylim = c(0, 3), height = 8,
width = 10, print_plot = FALSE, image_ext = ".png")
} # }