SMARTTR: A Mapping, Analysis, and Visualization Package for Wholebrain Dual-Ensemble Coronal Datasets.

SMARTTR. This package allows for the user-friendly pre-processing of segmentation data generated from ImageJ to a be compatible with the wholebrain package to generate region-based cell counts that are normalized by volume. It will also provides tools for data analysis based on experimental groupings.

Details

Object descriptions The data for analysis will be stored in objects that allow for more neat bundling of useful information together.

A slice object will contain all the data related to registration, segmentation for each channel, and cell counts for a particular image. It will also contain “metadata” about your experimental images, such as what the experimenter-assigned slice ID is, which brain atlas AP coordinate matches best with the given image, and what the path to the image used for registration is. These metadata are stored as the object’s attributes.

A mouse object is an object that will store multiple slice objects (and therefore all the information in it), and will eventually store the combined cell data and the region cell counts normalized by volume. Like a slice, it will also contain “metadata” about your mouse stored as attributes. An experimentAn experiment object consists of a list of processed mouse objects with raw data from slices omitted, and experimental attributes stored as a list. It will also contain “metadata” about your experimental personnel and analysis groups stored as attributes.

The package currently allows for easy implementation of the following steps

Setting up the pipeline by specifying experimentparameters, and save directories.

The interactive registration process.

Importing raw segmentation data from .txt files generated from ImageJ for multiple channels.

Optionally creating a filter for the 'cfos' and 'eyfp' channels to clean segmented counts.

Creating a segmentation object that is compatible with wholebrain functions.

Forward warping and mapping the data onto the standardized mouse atlas.

Cleaning the mapped data in all the following ways: + Removing cells that map outside the boundaries of the atlas.

Omitting regions by a default list of regions to omit.

Omitting regions by user specified region acronyms.

Removing Layer 1 cells

Removing cells from a contralateral hemisphere per slice if the registrations are divided by right and left hemispheres.

Obtaining cell counts normalized by region volume (per mm^2^) and region areas (per mm^2^).