This function performs a permutation analysis to compare the region pairwise correlation coefficients between two different analysis groups.

The data from two different analysis groups are compared by specifying the correlation_list_name_1 and correlation_list_name_2 parameters. Note that both of these analysis groups must have the same number of channels to compare. The functions get_correlations() needs to have been run for each of these analysis groups prior to running this function. The test statistics used is the pearson values of those in correlation_list_name_2 subtracted from corresponding Pearson values in correlation_list_name_1.

Usage

correlation_diff_permutation(
  e,
  correlation_list_name_1,
  correlation_list_name_2,
  channels = c("cfos", "eyfp", "colabel"),
  n_shuffle = 1000,
  method = "pearson",
  seed = 5,
  p_adjust_method = "none",
  alpha = 0.05,
  ...
)

Arguments

e

experiment object

correlation_list_name_1

(str) The name of the correlation list object used as the first group for comparison.

correlation_list_name_2

(str) The name of the correlation list object used as the second group for comparison.

channels

(str, default = c("cfos", "eyfp", "colabel")) The channels to process.

n_shuffle

(int, default = 1000) The number of permutation shuffles.

method

(str, default = "pearson", options = c("pearson", "spearman")) Specifies the type of correlations to compute. Spearman correlations are the Pearson linear correlations computed on the ranks of non-missing elements, using midranks for ties. See also Hmisc::rcorr()

seed

(int, default = 5) Random seed for future replication.

p_adjust_method

(bool or str, default = "none") Apply the named method to control for the inflated false discovery rate or FWER. Set to FALSE or "none" to keep "raw" p values. See also stats::p.adjust() for the correction options.

alpha

(float, default = 0.05) The alpha cutoff for significance between region pairwise correlation differences

...

additional parameters to pass to permute_corr_diff_distrib()

Value

e experiment object. The experiment object now has a list called permutation_p_matrix stored in it. Elements of this permutation_p_matrix list are the outputs of different permutation comparison analyses. These elements are named by the groups that were compared.

See also

Examples