get_registered_volumes (generic function)

Calculate the registered area (microns^2^) and the regional volumes (microns^3^) of all regions contained in a slice.

Note: Simplification of the analyzed regions by keywords is highly recommended because there are errors in the wholebrain basecode that results in a mismatch between the region acronym mapped to and the actual registration contour based on the region acronym. This mismatch is most notable in the dentate gyrus subregions. If simplification by keywords is used, this will circumvent the errors.

This function also automatically removes parent regions that are redundant, e.g. "CTX" should by volumetrically represented by summing all subregions, but there is a tiny amount of potential space that allows for cells to get mapped to slim spaces between subregions. This potential anatomical space should be ignored.

Calculate the registered area (microns^2^) and the regional volumes (microns^3^) of all regions contained in a slice. Note: Simplification of the analyzed regions by keywords is HIGHLY RECOMMENDED because there are errors in the wholebrain basecode that results in a mismatch between the region acronym mapped to and the actual registration contour based on the region acronym. This mismatch is most notable in the dentate gyrus subregions, where certain regions are represented twice because the DG curve along the rostral caudal axis. If simplification by keywords is used, this will circumvent the errors.

This function also automatically removes parent regions that are redundant, e.g. "CTX" should by volumetrically represented by summing all subregions, but there is a tiny amount of potential space that allows for cells to get mapped to slim spaces between subregions. This potential anatomical space should be ignored.

Usage

get_registered_volumes(x, ...)

# S3 method for slice
get_registered_volumes(
  x,
  simplify_regions = TRUE,
  simplify_keywords = c("layer", "part", "stratum", "division", "leaflet",
    "Subgeniculate", "island", "Islands", "Fields of Forel", "Cajal", "Darkschewitsch",
    "Precommissural"),
  ...
)

# S3 method for mouse
get_registered_volumes(
  x,
  slice_ID,
  hemisphere = NULL,
  simplify_regions = TRUE,
  simplify_keywords = c("layer", "part", "stratum", "division", "leaflet",
    "Subgeniculate", "island", "Islands", "Fields of Forel", "Cajal", "Darkschewitsch",
    "Precommissural"),
  replace = FALSE,
  ...
)

Arguments

x

a mouse or slice object

...

further arguments passed to or from other methods.

simplify_regions

(bool, default = TRUE ) simplify the normalized region counts based on keywords in the internal function, simplify_keywords

simplify_keywords

(str vec, default = c("layer","part","stratum","division")). Keywords to search through region names and simplify to parent structure

slice_ID

(str) ID of slice

hemisphere

(str) 'left', 'right' or NULL (both)

replace

(bool, default = FALSE). Replace previously calculated volumes for a particular slice.

Value

a mouse or slice object

s slice object with a stored dataframe with columns 'name' (full region name), 'acronym', 'area' (in microns^2^), 'volume' (in microns^3^) 'right.hemisphere'

m mouse object

Examples

if (FALSE) {
s <- get_registered_volumes(s)
}
if (FALSE) {
m <- get_registered_volumes(m, slice_ID = "1_10", hemisphere = "left", replace = FALSE)
}